initial attempt to add audio FX...didn't work, but wasn't silent

+++ Code/ChiCheBoard.cpp
@@ -1506,7 +1506,7 @@ void Board::Piece::Render( Board* board, bool highlight )
 	{
 		GLfloat specularity[] = { 1.f, 1.f, 1.f, 1.f };
 		GLfloat shininess[] = { 50.f };
-		GLfloat ambientDiffuse[] = { renderColor.get_e1(), renderColor.get_e2(), renderColor.get_e3(), 1.f };
+		GLfloat ambientDiffuse[] = { ( GLfloat )renderColor.get_e1(), ( GLfloat )renderColor.get_e2(), ( GLfloat )renderColor.get_e3(), 1.f };
 		GLfloat emissive[] = { 0.f, 0.f, 0.f, 0.f };
 
 		if( highlight )
